+++ sys/usb_board_reset/Kconfig
@@ -8,4 +8,4 @@
 config MODULE_USB_BOARD_RESET
     bool "Trigger a board reset via USB CDC ACM"
     depends on TEST_KCONFIG
-    depends on MODULE_USBUS_CDC_ACM || MODULE_STDIO_TINYUSB_CDC_ACM
+    depends on MODULE_USBUS_CDC_ACM || (MODULE_TINYUSB_DEVICE && MODULE_TINYUSB_CLASS_CDC)
